The memory accesses works on SPARC.
The problem is that SPARC required 32 memory accesses to be
on a 32 bit boundary. A set of macros GET_LONG and SET_LONG has been
written that splits longword fetchaes and stores into two 16 bit
accesses.

Theoretically, this would be required for 16-bit accesses as well (splitting
into two 8-bit accesses). However, only the 68030 ataris allow this. The older
68000-based machines gave bus error. It seems to be a reasonable performance
enhancement to require 16-bit aligned longword accesses for all emulated
architectures.

This can easily be changed if it is necessary for compatibility.